pyproject.toml
[project]
name = "gitlab-mcp"
version = "0.2.0"
description = "GitLab MCP (Model Context Protocol) server for interacting with GitLab API through AI Clients"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "mcp[cli]>=1.6.0",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "python-gitlab>=5.6.0",
    "asyncio>=3.4.3",
]

[project.optional-dependencies]
dev = [
    "black>=24.2.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "ruff>=0.11.5",
]
